chore: remove duplicate account controllers
This commit removes duplicate server controllers for accounts as
requested in #96.  All integration tests and end to end test pass,
indicating that this change removes unused code.

Closes #96.
